Add simple vue router structure
