+++ package.json
@@ -9,7 +9,8 @@
   },
   "dependencies": {
     "semantic-ui-css": "^2.4.1",
-    "vue": "^3.2.37"
+    "vue": "^3.2.37",
+    "vue-router": "^4.1.5"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.1.4",
