+++ src/pages/FavoritesPage.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>Favoritos</div>
+</template>
